[MOVE] : Friend 파일 위치 변경
