+++ app/src/main/java/com/sopt/now/presentation/home/friend/FriendAdapter.kt
@@ -5,7 +5,6 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import com.sopt.now.core.util.view.ItemDiffCallback
 import com.sopt.now.databinding.ItemFriendBinding
-import com.sopt.now.presentation.model.Friend
 
 class FriendAdapter() : ListAdapter<Friend, FriendViewHolder>(FriendAdapterDiffCallback) {
     override fun onCreateViewHolder(
